First official update of 2015
GreatShark brings Great Changes.
Hot on the heels of its release, I've already restyled the elements used
to highlight streams to be more aesthetically pleasing.
The link used to show permalinks is no longer a front row item, but will
slide into view as necessary.
The highlight colors used for the streams, selected journals, faves,
etc, and for notes is also more in line with the Vaunt Dark conceptual
view.

Welcome to the second GreatShark build.

Trivia: Did you know that the first build was finished and ready to be
committed when someone at FurAffinity pushed out another unannounced
change?
You do now. :)